+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix retrieval of memory limit and usage on Docker >= 20 (#394)
+
 ## [3.0.2] - 2025-12-22
 
 ### Fixed